clean up gateway lib by grouping rpc messaging module
