+++ gateway/ln-gateway/src/lib.rs
@@ -7,25 +7,20 @@ pub mod rpc;
 pub mod utils;
 pub mod webserver;
 
-use std::borrow::Cow;
-use std::collections::HashMap;
-use std::net::SocketAddr;
 use std::{
-    io::Cursor,
+    borrow::Cow,
+    collections::HashMap,
+    net::SocketAddr,
     sync::{Arc, Mutex},
     time::{Duration, Instant},
 };
 
 use axum::http::StatusCode;
 use axum::response::{IntoResponse, Response};
-use bitcoin::{Address, Transaction};
+use bitcoin::Address;
 use bitcoin_hashes::sha256::Hash as Sha256Hash;
 use fedimint_api::{Amount, NumPeers, TransactionId};
-use fedimint_server::{
-    config::ClientConfig,
-    modules::{ln::contracts::Preimage, wallet::txoproof::TxOutProof},
-};
-use futures::Future;
+use fedimint_server::{config::ClientConfig, modules::ln::contracts::Preimage};
 use mint_client::{
     api::{WsFederationApi, WsFederationConnect},
     ln::PayInvoicePayload,
@@ -35,141 +30,24 @@ use mint_client::{
 };
 use rand::thread_rng;
 use secp256k1::{KeyPair, PublicKey};
-use serde::{Deserialize, Deserializer, Serialize, Serializer};
 use thiserror::Error;
-use tokio::sync::{mpsc, oneshot};
+use tokio::sync::mpsc;
 use tracing::{debug, error, warn};
 use url::Url;
 
 use crate::{
     actor::GatewayActor,
     client::GatewayClientBuilder,
-    cln::HtlcAccepted,
     config::GatewayConfig,
     ln::{LightningError, LnRpc},
+    rpc::{
+        BalancePayload, DepositAddressPayload, DepositPayload, GatewayInfo, GatewayRequest,
+        InfoPayload, ReceivePaymentPayload, RegisterFedPayload, WithdrawPayload,
+    },
     webserver::run_webserver,
 };
 
 pub type Result<T> = std::result::Result<T, LnGatewayError>;
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct RegisterFedPayload {
-    pub connect: String,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct ReceivePaymentPayload {
-    // NOTE: On ReceivePayment signal from ln_rpc,
-    // we extract the relevant federation id from the accepted htlc
-    pub htlc_accepted: HtlcAccepted,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct InfoPayload;
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct BalancePayload {
-    pub federation_id: FederationId,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct DepositAddressPayload {
-    pub federation_id: FederationId,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct DepositPayload {
-    pub federation_id: FederationId,
-    pub txout_proof: TxOutProof,
-    #[serde(
-        deserialize_with = "serde_hex_deserialize",
-        serialize_with = "serde_hex_serialize"
-    )]
-    pub transaction: Transaction,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct WithdrawPayload {
-    pub federation_id: FederationId,
-    #[serde(with = "bitcoin::util::amount::serde::as_sat")]
-    pub amount: bitcoin::Amount,
-    pub address: Address,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct GatewayInfo {
-    pub version_hash: String,
-    pub federations: Vec<FederationId>,
-}
-
-#[derive(Debug)]
-pub enum GatewayRequest {
-    Info(GatewayRequestInner<InfoPayload>),
-    RegisterFederation(GatewayRequestInner<RegisterFedPayload>),
-    ReceivePayment(GatewayRequestInner<ReceivePaymentPayload>),
-    PayInvoice(GatewayRequestInner<PayInvoicePayload>),
-    Balance(GatewayRequestInner<BalancePayload>),
-    DepositAddress(GatewayRequestInner<DepositAddressPayload>),
-    Deposit(GatewayRequestInner<DepositPayload>),
-    Withdraw(GatewayRequestInner<WithdrawPayload>),
-}
-
-#[derive(Debug)]
-pub struct GatewayRequestInner<R: GatewayRequestTrait> {
-    request: R,
-    sender: oneshot::Sender<Result<R::Response>>,
-}
-
-pub trait GatewayRequestTrait {
-    type Response;
-
-    fn to_enum(self, sender: oneshot::Sender<Result<Self::Response>>) -> GatewayRequest;
-}
-
-macro_rules! impl_gateway_request_trait {
-    ($req:ty, $res:ty, $variant:expr) => {
-        impl GatewayRequestTrait for $req {
-            type Response = $res;
-            fn to_enum(self, sender: oneshot::Sender<Result<Self::Response>>) -> GatewayRequest {
-                $variant(GatewayRequestInner {
-                    request: self,
-                    sender,
-                })
-            }
-        }
-    };
-}
-
-impl_gateway_request_trait!(InfoPayload, GatewayInfo, GatewayRequest::Info);
-impl_gateway_request_trait!(RegisterFedPayload, (), GatewayRequest::RegisterFederation);
-impl_gateway_request_trait!(
-    ReceivePaymentPayload,
-    Preimage,
-    GatewayRequest::ReceivePayment
-);
-impl_gateway_request_trait!(PayInvoicePayload, (), GatewayRequest::PayInvoice);
-impl_gateway_request_trait!(BalancePayload, Amount, GatewayRequest::Balance);
-impl_gateway_request_trait!(
-    DepositAddressPayload,
-    Address,
-    GatewayRequest::DepositAddress
-);
-impl_gateway_request_trait!(DepositPayload, TransactionId, GatewayRequest::Deposit);
-impl_gateway_request_trait!(WithdrawPayload, TransactionId, GatewayRequest::Withdraw);
-
-impl<T> GatewayRequestInner<T>
-where
-    T: GatewayRequestTrait,
-    T::Response: std::fmt::Debug,
-{
-    async fn handle<F: Fn(T) -> FF, FF: Future<Output = Result<T::Response>>>(self, handler: F) {
-        let result = handler(self.request).await;
-        if self.sender.send(result).is_err() {
-            // TODO: figure out how to log the result
-            tracing::error!("Plugin hung up");
-        }
-    }
-}
 
 pub struct LnGateway {
     config: GatewayConfig,
@@ -445,35 +323,6 @@ pub enum LnGatewayError {
     UnknownFederation,
     #[error("Other: {0:?}")]
     Other(#[from] anyhow::Error),
-}
-
-pub fn serde_hex_deserialize<'d, T: bitcoin::consensus::Decodable, D: Deserializer<'d>>(
-    d: D,
-) -> std::result::Result<T, D::Error> {
-    if d.is_human_readable() {
-        let bytes = hex::decode::<String>(Deserialize::deserialize(d)?)
-            .map_err(serde::de::Error::custom)?;
-        T::consensus_decode(&mut Cursor::new(&bytes))
-            .map_err(|e| serde::de::Error::custom(format!("{:?}", e)))
-    } else {
-        let bytes: Vec<u8> = Deserialize::deserialize(d)?;
-        T::consensus_decode(&mut Cursor::new(&bytes))
-            .map_err(|e| serde::de::Error::custom(format!("{:?}", e)))
-    }
-}
-
-pub fn serde_hex_serialize<T: bitcoin::consensus::Encodable, S: Serializer>(
-    t: &T,
-    s: S,
-) -> std::result::Result<S::Ok, S::Error> {
-    let mut bytes = vec![];
-    T::consensus_encode(t, &mut bytes).map_err(serde::ser::Error::custom)?;
-
-    if s.is_human_readable() {
-        s.serialize_str(&hex::encode(bytes))
-    } else {
-        s.serialize_bytes(&bytes)
-    }
 }
 
 impl IntoResponse for LnGatewayError {
